// UI/src/enums/enums.ts
export enum Note {
    A = "A",
    ASharp = "ASharp",
    B = "B",
    C = "C",
    CSharp = "CSharp",
    D = "D",
    DSharp = "DSharp",
    E = "E",
    F = "F",
    FSharp = "FSharp",
    G = "G",
    GSharp = "GSharp"
}

export enum Scale {
    Major = "Major",
    Minor = "Minor",
    HarmonicMinor = "HarmonicMinor",
    MelodicMinor = "MelodicMinor",
    PentatonicMajor = "PentatonicMajor",
    PentatonicMinor = "PentatonicMinor",
    Blues = "Blues",
    Dorian = "Dorian",
    Phrygian = "Phrygian",
    Lydian = "Lydian",
    Mixolydian = "Mixolydian",
    Locrian = "Locrian"
}